[refactor/#73] type 패키지 이동 및 아이콘 리소스 추가
