--- app/src/main/java/com/smashing/app/core/common/type/DialogType.kt
@@ -1,7 +1,0 @@
-package com.smashing.app.core.common.type
-
-enum class DialogType {
-    ALERT,
-    CONFIRM;
-}
-
