+++ app/src/main/java/com/smashing/app/core/designsystem/component/badge/TierBadge.kt
@@ -17,8 +17,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.smashing.app.core.common.type.TierType
-import com.smashing.app.core.common.type.TierType.*
+import com.smashing.app.data.type.TierType
+import com.smashing.app.data.type.TierType.*
 import com.smashing.app.core.designsystem.theme.SmashingAndroidTheme
 import com.smashing.app.core.designsystem.theme.SmashingTheme.colors
 import com.smashing.app.core.designsystem.theme.SmashingTheme.typography
