+++ app/src/main/java/com/smashing/app/core/designsystem/component/card/MatchingCard.kt
@@ -25,8 +25,8 @@ import com.smashing.app.R
 import com.smashing.app.R.drawable.ic_close_sm
 import com.smashing.app.R.drawable.ic_link
 import com.smashing.app.R.string.accept
-import com.smashing.app.core.common.type.GenderType
-import com.smashing.app.core.common.type.TierType
+import com.smashing.app.data.type.GenderType
+import com.smashing.app.data.type.TierType
 import com.smashing.app.core.designsystem.component.button.SmashingBaseButton
 import com.smashing.app.core.designsystem.state.MatchingCardState
 import com.smashing.app.core.designsystem.state.MatchingCardState.Confirm
