+++ app/src/main/java/com/smashing/app/core/designsystem/component/sport/SportSkillSelector.kt
@@ -11,10 +11,6 @@ import androidx.compose.foundation.selection.selectableGroup
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
@@ -31,7 +27,7 @@ import com.smashing.app.core.designsystem.theme.SmashingTheme.colors
 import com.smashing.app.core.designsystem.theme.SmashingTheme.typography
 import com.smashing.app.core.extension.noRippleClickable
 import com.smashing.app.presentation.signup.component.SignUpTitle
-import com.smashing.app.core.common.type.SkillType
+import com.smashing.app.data.type.SkillType
 import kotlinx.collections.immutable.toImmutableList
 
 @Composable
