+++ app/src/main/java/com/smashing/app/data/model/matching/ReceivedMatching.kt
@@ -1,8 +1,7 @@
 package com.smashing.app.data.model.matching
 
-import com.smashing.app.core.common.type.GenderType
-import com.smashing.app.core.common.type.SportType
-import com.smashing.app.core.common.type.TierType
+import com.smashing.app.data.type.GenderType
+import com.smashing.app.data.type.TierType
 
 data class ReceivedMatching(
     val matchingId: String,
