+++ app/src/main/java/com/smashing/app/data/model/profile/ActiveUserProfile.kt
@@ -1,7 +1,7 @@
 package com.smashing.app.data.model.profile
 
-import com.smashing.app.core.common.type.SportType
-import com.smashing.app.core.common.type.TierType
+import com.smashing.app.data.type.SportType
+import com.smashing.app.data.type.TierType
 
 //TODO nickname과 region은 서버에서 추가 예정
 data class ActiveUserProfile(
