+++ app/src/main/java/com/smashing/app/presentation/matching/MatchingScreen.kt
@@ -90,6 +90,7 @@ private fun MatchingScreen(
     Column(
         modifier = modifier
             .fillMaxSize()
+            .background(color = SmashingTheme.colors.bgCanvas)
             .padding(horizontal = 16.dp),
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
