+++ app/src/main/java/com/smashing/app/presentation/profile/ProfileContract.kt
@@ -1,8 +1,8 @@
 package com.smashing.app.presentation.profile
 
 import androidx.compose.runtime.Immutable
-import com.smashing.app.core.common.type.SportType
-import com.smashing.app.core.common.type.TierType
+import com.smashing.app.data.type.SportType
+import com.smashing.app.data.type.TierType
 import com.smashing.app.data.model.profile.RatingCount
 import com.smashing.app.data.model.profile.Review
 import com.smashing.app.data.model.profile.TagCount
