+++ app/src/main/java/com/smashing/app/presentation/profile/user/copmponent/ProfileTierBox.kt
@@ -30,8 +30,8 @@ import com.smashing.app.R.drawable.ic_plus
 import com.smashing.app.R.string.lp_remaining_text
 import com.smashing.app.R.string.lp_status
 import com.smashing.app.R.string.tier_description
-import com.smashing.app.core.common.type.SportType
-import com.smashing.app.core.common.type.TierType
+import com.smashing.app.data.type.SportType
+import com.smashing.app.data.type.TierType
 import com.smashing.app.core.designsystem.component.badge.TierBadge
 import com.smashing.app.core.designsystem.component.button.SmashingBaseButton
 import com.smashing.app.core.designsystem.component.chip.SmashingChip
@@ -180,5 +180,3 @@ private fun ProfileTierBoxPreview() {
         )
     }
 }
-
-
