+++ app/src/main/java/com/smashing/app/presentation/profile/user/copmponent/UserProfileCard.kt
@@ -2,7 +2,6 @@ package com.smashing.app.presentation.profile.user.copmponent
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -28,8 +27,8 @@ import com.smashing.app.R.string.record_label
 import com.smashing.app.R.string.review
 import com.smashing.app.R.drawable.ic_man_20
 import com.smashing.app.R.drawable.ic_woman_20
-import com.smashing.app.core.common.type.GenderType
-import com.smashing.app.core.common.type.TierType
+import com.smashing.app.data.type.GenderType
+import com.smashing.app.data.type.TierType
 import com.smashing.app.core.designsystem.component.badge.TierBadge
 import com.smashing.app.core.designsystem.component.image.UrlImage
 import com.smashing.app.core.designsystem.theme.SmashingAndroidTheme
