+++ app/src/main/java/com/smashing/app/presentation/signup/SignUpContract.kt
@@ -1,9 +1,9 @@
 package com.smashing.app.presentation.signup
 
 import androidx.compose.runtime.Immutable
-import com.smashing.app.core.common.type.GenderType
-import com.smashing.app.core.common.type.SkillType
-import com.smashing.app.core.common.type.SportType
+import com.smashing.app.data.type.GenderType
+import com.smashing.app.data.type.SkillType
+import com.smashing.app.data.type.SportType
 
 interface SignUpContract {
     @Immutable
